implement detailed log table

+++ app/src/main/java/com/taurilogs/app/api/RequestParams.kt
@@ -6,6 +6,6 @@ class RequestParams(
     val r: String?, // realm
     val cn: String?, // character name
     val limit: Int?, // limit
-    val id: Int?, // log id
+    val id: Long?, // log id
 ) {
 }
